refactor: rename all.py to combined.py (avoid shadowing builtin)
